added a small description of project to readme, created folders: validators and controllers to make code more clean
