+++ controllers/contacts/indexContacts.js
@@ -1,0 +1,8 @@
+const { listContacts } = require("../../models/contacts");
+
+async function indexContacts(req, res, next) {
+  const contacts = await listContacts();
+  return res.json(contacts);
+}
+
+module.exports = { indexContacts };
