+++ models/contacts.json
@@ -1,11 +1,5 @@
 [
   {
-    "id": "qdggE76Jtbfd9eWJHrssH",
-    "name": "Ally Raymondy123",
-    "email": "123@wp.pl",
-    "phone": "(992) 914-37922123"
-  },
-  {
     "id": "drsAJ4SHPYqZeG-83QTVW",
     "name": "Kennedy Lane",
     "email": "mattis.Cras@nonenimMauris.net",
